chore(java_extraction): move ForwardingFileManager to a better package
