+++ kythe/java/com/google/devtools/kythe/platform/java/filemanager/ForwardingStandardJavaFileManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.devtools.kythe.common;
+package com.google.devtools.kythe.platform.java.filemanager;
 
 import com.google.common.flogger.FluentLogger;
 import java.io.File;
